Added some much needed restructuring to the scene and game object code. It still needs more work, but should be easier to use from a pure c++ standpoint.
Updated utTypes.h to handle internal iterators.
Fixed utHashTable bugs and improved overall performance.

Added a simple utHashSet as a replacement for std::set.

+++ Engine/Utils/utRandom.cpp
@@ -140,7 +140,7 @@ float utRandomNumberGenerator::randNormal(float mean, float deviation)
 		b = randRange(-1.0f, 1.0f);
 		s= a*a+b*b;
 	}
-	b = a * sqrt((-2.0*log(s)) / s);
+	b = a * (float)sqrt((-2.0*log(s)) / s);
 	return mean+ b*dev;
 }
